requirements/common.txt
transformers==4.52.4
torch==2.7.1
datasets
pydantic
hydra-core
accelerate==1.8.1
lm-eval==0.4.9
loguru
python-dotenv
math_verify
sympy